RF: distributions: Add "walk up path, find root" helper
This class extracts logic from CondaTracer._get_conda_path.  That
method takes a path and walks up its parent directory until it finds
something that looks like the root of a conda environment.  Also, it
caches path->root mappings for later runs.

Move this walk/cache part to a separate function because we'll use
this pattern when identifying virtualenv roots.