RF: conda: Tweak code for retrieving pip dependencies
There is no need to .get() rather than __getitem__ because the
previous line checks that the "pip" key was present.  Also, we only
expect one pip entry, so break after we find it.  (And if we did
expect more than one, the previous code should be changed to update
the list rather than overriding it.)