TST: Move VCS tests to distributions/tests/test_vcs.py
For consistency with tests for Conda and Debian tracers, move the VCS
tracer parts from interface/tests/test_retrace_packages.py to
distributions/tests/test_vcs.py.

All of the removed parts from test_retrace_packages should now be
covered in test_vcs.  In addition, because we are now using the
fixtures, the url/pushurl test shouldn't be a noop on Travis [*].

[*] https://github.com/ReproNim/niceman/pull/165#discussion_r163300427