BF: vcs: Fix the contains=True check
We get the remotes that contain the hexsha with "git branch -r
--contains=".  However, we don't actually split the lines of the
output, so the list of containing remotes we get is just a unique set
of the characters in the git output string.

Update the code to actually split the lines.  While we're here, switch
the call over to for-each-ref plumbing.

Fixes #166.