BF: vcs: Replace 'lstrip' with 'strip' for older Git version
The 'lstrip' for specifier for 'git for-each-ref' wasn't introduced
until Git v2.13.0.  There is an older synonym for it, 'strip', which
Git still retains so that it doesn't break third-party scripts.

This probably explains why the test_git_repo_remotes tests pass on
Travis yet a failure was reported in #182.