BF: vcs: Don't expect the 'git for-each-ref' call to fail
With expect_fail=True, this will return None, so the splitlines call
will fail with an AttributeError.  But, at this point, we know we're
in a Git repository (we have a hash after all), so this command
shouldn't fail.  If no remote branches contain the hash, an empty
string should be returned, so we're safe to split it.

Re: #182