RF: piputils: Add a 'list' function that keeps other fields
get_pip_packages drops the version and location because we get these
from pip show.  However, the location from pip show and pip list is
different: pip list only shows it for editable packages, whereas pip
shows the location for non-editable packages as well.

We'll use this difference to flag a pip package as editable, so add a
function that preserves this information.