BF+ENH: conda/venv: Flag editable pip packages
Before 6762a746, the conda tracer used "origin_location" to refer to
the location for editable packages _only_.  Since then, the conda
tracer and the venv tracer have conflated the origin location with pip
show's "Location" field, where the location is for both editable and
non-editable packages.

Restore this distinction by adding an "editable" attribute to
VenvPackage and CondaPackage.  Use the fact that the output of 'pip
list' only contains the location for editable packages to determine
whether a package is editable.

Rename "origin_location" to "location" since it no longer is meant to
be restricted to locations for editable packages.

Note that, for conda, we make a separate call to 'pip list' rather
than relying on the pip packages in the 'conda env export'
output (which are generated by a pip list call underneath) because
'conda env export' calls 'pip list' in a way that can drop the
location for editable packages.  (See 54da3757 and #163.)