ENH: conda: Extend unknown files with editable pip packages
This is a followup to #181, which modified the tracers to do multiple
passes on the set of unknown files.  The venv tracer is currently the
only tracer that will extend the set of unknown files that it passes
back.  It does so if it encounters a pip package with a location that
is outside of the virtualenv path, which happens in the case of
system-wide installs as well as editable packages.

Make conda add editable pip packages to the unknown files.  Don't
worry about system-wide packages because they aren't exposed in conda
environments.

Closes #189.