BF: Add kludge to load Git and SVN distributions
NicemanProvenance is unable to load a spec that includes a
GitDistribution or SVNDistribution because these don't follow the
expected naming conventions.  Specifically, GitDistribution and
SVNDistribution are in the module "vcs" rather than "git" and "svn",
respectively, and the "SVN" in SVNDistribution is all uppercase.

We could (1) adjust Distribution.factory to treat SVN and Git as
special cases, (2) restructure things so that the SVN and Git follow
the expected naming pattern, or (3) maintain a mapping from
distribution["name"] -> module.class.

Go with (1) for now because it involves the fewest changes.

Fixes #222.