ENH: debian: Make the dpkg "multiple packages" warning stricter
Depending on how specific the pattern passed to 'dpkg-query -S' is,
the output may contain hits where the path is a general
directory (e.g., /usr/lib/python3.5) that maps to *many* packages.

This leads to warning messages that flood the screen about something
that isn't problematic, so don't issue a warning if the path is a
directory.  Because Debian sets HANDLES_DIRS = False, we could just
filter out directories regardless of the "," in line check, but that
would mean doing many more isdir calls.

Fixes #224.