RF: Replace remaining attr.ib calls with attrib helper
These are calls to attr.ib that do not specify the `default` argument.
Replace these with the more explicit `attrib(default=attr.NOTHING,
...)` so that we consistently use attrib across the codebase.
Although this is more verbose, the consistency is important because
otherwise the reader has to detect the one character difference
between "attr.ib" and "attrib" in order to know what the default
behavior is.