NF: Add initial support for "installing" Git repositories
This adds basic support for instantiating Git repositories listed in a
GitDistribution object.  If a repository isn't present, we clone it
from one of the recorded sources.  Otherwise, if the recorded revision
isn't present, we fetch a remote that is marked as containing that
revision.  GitDistribution.install_packages' main goal is to check out
the same revision, but it also sets up the recorded remotes and checks
outs the recorded branch if possible.