CLN: Purge CRLF line breaks
This is the result of running dos2unix on the tracked files.  In
general we might want to avoid this sort of widespread churn, but in
this case all the files are __init__.py files that lack content worth
blaming and that are unlikely to cause conflicts with topic branches.