RF: Absorb more resource handling into ResourceManager
Incorporate parts of @yarikoptic's gh-98 (yarikoptic/rf-manager),
focusing on only the changes that concerned the ResourceManager.  Like
that PR, reduce the process of obtaining a resource to a single method
call, handling the inventory behind the scenes, and break up
get_resource_info into more methods.

Also, like that PR, rework resource name/ID handling.  Try to infer
whether the resource references a name or ID, and, for cases where a
reference is ambiguous, expose a flag that enables specifying whether
a reference is a name or ID.  Each resource reference is consistently
called a "resref" in the code to avoid confusion with a resource
instance.

The interface tests have become a little more cumbersome because we
now have to mock an instance rather than the class.  For clarity and
maintainability, it's probably worth exploring whether we can make a
fixture or some other wrapper that simplifies the process.