ENH+RF: Add --config as a global option
This replaces the individual --config options that a previous commit
in this series removed from many interface commands.

Adjust all the interfaces to use a helper function instead of directly
importing the ResourceManager instance so that we can delay the
instance creation long enough for any files specified via --config to
have an effect.