DOC: config: Remove mention of $PWD config
This documentation has always been technically wrong in the sense that
ConfigManager has never actually looked for a file directly in $PWD.
But resource/base.py did add $PWD/niceman.cfg to the list of config
files, so it was effectively true.  As of f221a199 (RF: resource:
Don't look for config file in the current directory, 2018-07-31),
that's no longer true.