BF: resources: Avoid trying to create directory for empty string
Like the code that it replaced, _prepare_dest_path() does not check
that there's a leading directory before calling mkdir on an empty
string.  Add an explicit check for a leading path.

Note: This worked by chance for some resources because their exists()
methods return true for an empty string.  This should be fixed
separately.