BF: Restrict singularity and docker put() to absolute targets
The put() docstring doesn't explicitly mention whether the destination
is expected to be an absolute path, but based on the put() call in
test_ssh_class and on that fact that there is no explicit isabs()
check in any put() method, relative paths seem to be allowed.

For ssh sessions, there is an obvious default working directory: the
user's home directory. For local shell's, the put() is just a get(),
so it defaults to the current directory.  The situation isn't as clear
for docker and singularity, and their methods don't currently provide
any special handling.  The docker put() method fails if there is no
leading directory, and it puts the files in "/" if there is a
non-absolute leading path.  The singularity put() method appears to
use a path that matches the local directory if it exists in the
singularity image and a /home/USER otherwise.

Adjust singularity and docker to signal an error when a non-absolute
path is given, at least until these methods are updated with explicit
rules for handling non-absolute paths.