ENH: Don't list resources if loading a known resource fails
Usually importing a resource fails because the user has passed an
unknown resource type, so we report the known resources in the
exception message.  But if the import of a known resource fails (e.g.,
because the resource imports an uninstalled module), it's confusing to
show a list of known resources.

Closes #297.