BF+ENH: session.{exists,isdir}_command: Escape path
Most the session methods don't require path escaping because they pass
the path as an item of the command list, but exists_command() and
isdir_command() format the string as part of "bash -c ...", so they
need to escape the path.

This will be tested in the upcoming changes to
test_session_abstract_methods.