BF: singularity: Don't convert string commands to a list
There are two issues with the current conversion from string to
list: (1) we use a plain str.split() but the string may contain
special escaping/quoting and (2) the string command may contain things
like redirect.  The first issue could be fixed by using shlex.split(),
but we need to keep the command in string form to handle the second
issue.  So, instead, convert the list form to the string form.

This will be tested in the upcoming changes to
test_session_abstract_methods.