BF: redhat: Don't instantiate RPMSource with unknown fields
test_redhat.py::test_tracer is failing on Travis with

    TypeError: __init__() got an unexpected keyword argument 'excluded'

Presumably this is due to a change in the Docker image used for this
test.

Instead of dying in such a situation, issue a warning that we've
encountered an unknown field in the 'yum --verbose repolist' output
and continue.  Perhaps we should add an "excluded" attributes to
RPMSource and/or update the test to use a stable Docker image, but we
should still make _find_all_sources resilient to unknown keys to avoid
such failures in the future.