CLN+TST: Purge print statements in tests
All of these appear to be meant manual debugging and don't affect the
actual tests.  Remove them so that they don't pollute the test output.
This becomes more important as of 3ef0f5c0 (Added -s switch for pytest
in travis, 2018-09-13) because the test stdout/stderr is dumped to the
Travis logs, making the logs exceed Travis's online display limit.

Closes #322.