RF: execute: Rework different command variants into a common interface
execute can handle three command "types": (1) an "internal" command
for Session.niceman_exec, (2) a plain command for
Session.execute_command, and (3) a tracer-wrapped command for
Session.execute_command.  Orchestrating the tracing of course is only
needed for the last variant, but a large chunk of Execute.__call__ is
concerned with it.

Make an adapter for all these variants so that Execute.__call__ can
treat them all the same.  The traced command code is still ugly, but
at least it's confined to a single class with at least somewhat
informative method names.  In addition to making the shared parts of
Execute.__call__ more obvious, this makes testing the trace code more
straightforward and makes it easier to move the trace functionality to
the upcoming `run` command, if we decide it's better there.