BF: retrace: Let Python callers pass a session instance
Before 309bdbbe (RF: Absorb more resource handling into
ResourceManager, 2018-07-27), retrace accepted a Session instance as
its `resource` argument.  That was a bit confusing because it was
undocumented and because, despite the parameter name, a resource
object wasn't accepted.  (The previous code was also buggy because a
CLI caller couldn't specify a resource).  But `execute --trace`
expects to be able to pass a session instance to `retrace`, so
309bdbbe introduced a regression by removing that handling.

Passing session instances may be something we want to allow more
generally in the Python API, but for now just add a kludge to
`retrace` so that a Session instance can be passed for `resref`.

Fixes #316.