ENH: orchestrators: Add status
Submitters already have a status property that is determined by
querying the batch system.  But the runscript also updates the file
META_DIRECTORY/status with status information.  In many situations,
this value is more meaningful because it will persist even after the
batch system forgets about the job.  For example, after a successful
job, META_DIRECTORY/status is "completed", but the status determined
from the batch system may be "unknown".

Add a method the the Orchestrator class that looks at
META_DIRECTORY/status.  We probably want to rethink how these
different statuses interact and how we should access them, but for now
just override the "our status" component of Submitter.status with the
value from META_DIRECTORY/status so that we can use it when listing
unfetched jobs.