RF: orchestrators: Drop _{working,root}_directory attributes
We don't need them with the cached_property decorator.  We no longer
pop root_directory and working_directory, so the template
instantiation requires a bit of special handling so that we don't pass
duplicate keyword arguments.  The duplicate values should never be
different, but still make sure that the property values take
precedence since those values are the ones the orchestrator actually
uses.