ENH: orchestrators: Add simple, non-DataLad variant
PlainOrchestrator may not be too useful on its own because we'd
recommend users have a dataset locally, but it still is worth adding
so that there's an orchestrator that doesn't require DataLad.  Plus,
PrepareRemotePlainMixin will be used to add a "datalad-run"
orchestrator variant that doesn't depend on having DataLad remotely.