ENH: job templates: Create niceman ref for "datalad-pair"
We already to this for the "datalad-pair-run".  These refs were more
useful when "datalad-pair-run" use to do a hard reset at the end, but
even without that they're still helpful for inspecting/debugging the
remote repository.  Also, once we allow checking out the target
commit, a run may not be on the mainline, so creating a ref means
those runs aren't detached.