BF: orchestrators: Check for existing remote
The remote may already exist even when the remote dataset doesn't if
the dataset was deleted on the remote or if another remote already
happens to have that name.  Either way, creating the sibling will
fail, so give a better error.

It's still possible that the remote dataset exists but the git remote
is for another repo/resource.  To cover that, we'd have to compare the
remote URLs, so punt on that for now.