ENH: orchestrators: Check out target commit
Make sure the remote dataset is at the right commit before executing
the command.

This adds one of the main setup steps we lack when preparing the
remote dataset, but there are still issues with it.  The biggest is
that detaching to a commit doesn't work well with the DataLad update()
call, since the changes won't be fetched.  If we could pass arbitrary
refs to update(), we could tell it to fetch the refs/niceman/<jobid>
ref.  Or, if we just wanted to solved this for the "run" variant, we
could fetch that ref ourselves, since the content handling is already
taken care of with the post-run tarball.