TST: execute: Fix skipif condition
There are two issues with this condition: (1) it was intended to say
"not in" rather than "in" and (2) an ExternalVersions instance's
__contains__(item) always returns False if __getitem__(item) hasn't
been called previously.  The combination of no. 1 and no. 2 means the
test would never be skipped.

No. 2 has been fixed as of DataLad's e71a57aa3 (BF: external_versions:
Make __contains__ work with non-primed items, 2018-12-03) and has been
incorporated via 3rd.

Update the skipif condition accordingly.