ENH: orchestrator: Detect an unprocessed job
We've got four sources of status information: (1) what we query from
the batch system, (2) our normalization of no. 1 into a common set of
statuses, and (3) the status from the metadata status file that the
runscript updates during the job, and (4) the sentinel file that is
touched when post-processing completes.

We use no. 2 to decide whether the batch system is done running the
job.  Because Submitter.follow() doesn't look at (and isn't even aware
of no. 4), we don't signal that something is wrong if the
post-processing is incomplete because post-processing failed.

Rework Orchestrator's status attribute and give it a follow() method
that wraps Submitter.follow(), looking at no. 4 to see if
post-processing succeeded.