BF: orchestrators: Relocate "processing complete" sentinel
eb0a8c683 (ENH: job templates: Use separate file for final status,
2018-12-06) started using a separate file for the "processing
complete" sentinel.  This leaves an untracked file in the repo, though,
so it makes the "dirty dataset" check for any subsequent runs fail.

Touching the sentinel before the "save repo" step would fix this
issue, but we want the touch to be the very last step to try to avoid
labeling the processing as complete despite there actually being a
failure.  Instead put this file outside of the repo.