ENH: orchestrators: Support "container" parameter
Teach the DataLad-based orchestrators about dataset containers
configured via the "datalad-container" extension.  This boils down to
extracting the appropriate values from the configuration and then
tweaking the command string.

For singularity images, this approach will work even if the remote
resource does not have the datalad-container extension installed.  For
docker images that are used via datalad-container's adapter, this
won't work because (1) the adapter isn't installed unless
datalad-container is and (2) the user needs permissions to run "docker
load" on the resource.

At the moment, this job_spec key isn't advertised.  We need to come up
with a general way to list the available job spec parameters.