ENH: orchestrators: Try to get public data before copying
DataLad-based orchestrators copy inputs via 'datalad publish'.  Update
them to first try to get the publicly shared inputs and then if needed
follow up with a second call to publish. We could be smarter
here (e.g., checking which inputs have URL or looking at what's
present locally), but this dumb approach works for now and has the
advantage that it should work across dataset boundaries.

This behavior should probably be configurable from (at least) the job
spec, but punt on that for now.