ENH: orchestrators: Add message parameter
This should do something useful for any orchestrator that commits, but
the main motivation is to be able to give custom message to the
datalad-run based orchestrators.

Add a parameter to run's interface (in addition to -b message= or the
message job spec key) in hopes of encouraging the use of custom
messages.