BF: orchestrators: Fix regression in formatting SSH URL
83af70bd9 (ENH: orchestrators: Towards supporting AWS, 2019-01-04)
made it possible to format an SSH URL from an AWS resource, but it
broke the formatting for standard SSH resources.  In order to work
with SSH config settings, we need to use resource.host, which can
match a host nickname from .ssh/config, not the connection's host,
which is the real host name.