BF: virtualenv: Don't return internal symlinks as unknown paths
We resolve virtualenv symlinks and report them as unknown files
because links in the virtual environment may point to system
packages ($ENV/lib/python3.5/types.py -> /usr/lib/python3.5/types.py),
and we want to detect those resolved paths with another
tracer (usually DebTracer) on the next `retrace` iteration.

However, a symlink may point to another file under the same virtual
environment, as is the case with "$ENV/bin/python -> $ENV/bin/pythonX"
links.  Returning the resolved path of an internal link as an unknown
file to `retrace` will result in a duplicate [*] virtualenv
distribution because the path will reach the virtualenv tracer on the
next iteration.  (This can in turn lead to duplicate Git distributions
if any of the virtualenv packages are editable.)

Re: #367

[*]: Or at least near-duplicate: the second distribution will lack any
     files that were assigned to virtualenv packages during the first
     pass.