BF: orchestrators: Honor job spec's local_directory
jobs() resurrects orchestrators after switching to the recorded local
directory because some of the setup assumes the current directory
matches the original current directory.  However, the local_directory
property may be accessed after this context is exited, so it should
look at the job spec before falling backing to os.getcwd().  (Note:
The DataLad orchestators don't have this issue because the overload
local_path to be ds.path.)