RF: orchestrators: Update DataLad version check
The required feature comes after DataLad v0.11.2, and v0.11.3 isn't
yet released.  Checking whether the DataLad version is greater than
0.11.2 makes it possible to test this functionality with the
developmental version of DataLad.  But it isn't a check we should keep
around because the user may be on a revision after v0.11.2 but before
the feature was added.  DataLad v0.11.3 should be released soon, so
use the stricter "at least 0.11.3 check".