RF: singularity: Don't store runner as attr
We assume that connect() will be called first and that it will set the
runner attribute, but that undocumented assumption is prone to
breakage and there isn't a good reason to use an attrib here.  Instead
do what we do in SingularitySession: initialize a class attribute to a
runner instance.