TST+NF: skip: New module for registering skip conditions
Most of the skip functions in tests.utils provide the ability to skip
a function using a decorator or an inline call.  They do this by
behaving as a decorator if the function argument is provided.  The
main issue with this approach is that it requires adding very similar
boilerplate for each new skip_* function.

Provide a way to avoid this boilerplate.  With this new module, both
the decorator and inline variants can be produced by defining a
condition function and adding it to a list.