RF+TST: session: Rework fixture to allow more fine-grained skips
The test_session module is written in a way that assumes our optional
dependencies are available.  We can't just import the resource modules
late and skip some tests because the resource_session fixture contains
the logic for all of the different session classes.  This means that,
for test_session_abstract_methods, it's an all-or-none skip.

To work around this, remove the resource_session fixture.  Isolate all
of its "check logic" into a new fixture that produces a check
function.  The only changes to this check code (aside from
indentation) are that the isinstance() conditions are replaced with
string comparisons because this new fixture cannot import the session
objects (that'd put us back in the same situation).

Replace test_session_abstract_methods with separate test functions for
each set of "dependencies" (none, ssh/docker, docker).