TST: skip: Add condition that checks if boto3 is installed
We'll use this to adjust our test suite to work when optional packages
aren't installed.  Use "aws" rather than "boto3" in the name because
that better captures the intent of the test and it means we only have
to adjust the condition, not all the callers, if our AWS dependencies
change.