TST: travis: Remove BOTO_CONFIG=/tmp/nowhere
This came from the initial import of DataLad code.  The intention
seems to be to set this variable for _all_ runs [0].  However, we have
always set this a _matrix_ variable [1], which results in a single
run.

So, _if_ we set this, we should set this under "env: global:".  But
the other runs that don't set this variable have been passing, so the
original Travis issue that prompted BOTO_CONFIG=/tmp/nowhere must not
be present anymore.  Remove the run.

[0] Since DataLad's 838e269d7 (BF: in blind hope reincarnating
    BOTO_CONFIG=/tmp/nowhere, 2016-09-13), it sets this variable is
    under "env: global: ...".

[1] The last commit changed put it directly under "env: matrix:", but
    but it used be directly under "env:".  In that case matrix is
    implied, so the effect is the same.