TST: travis: Collapse 3.4 and core deps runs into a single run
Combining these isn't ideal because, if this run fails, it's not as
obvious if the problem is that we did something incompatible with
py3.4 or that we are assuming an optional dependency is always
present.  But, in most cases, it shouldn't take much effort to discern
between these two possibilities, and this gets us down to five
runs (which seems to be Travis's default limit for concurrent jobs?).