ENH: retrace: Try mapping pyc files back to their source
When pyc files come in, convert them to the non-cached name because
the tracers (or at least VenvTracer and DebTracer) do not map pycache
files to packages.  The main decision is _where_ to do this.  Instead
of doing the pyc -> py mapping in all the tracers, do it once at the
retrace entrypoint.