ENH: virtualenv: Resolve symlink directories too
If a symlink file under the virtual environment points outside the
environment, we resolve it and replace the original path in the
unknown files set with the resolved path.  This allows other tracers
to detect that file as part of one of their packages.

But virtual environments can also contained linked directories, such
as

  $VENV/lib/python3.5/collections/ -> /usr/lib/python3.5/collections/

To handle these, unconditionally call realpath() on subpaths of the
virtual environment.