RF: ls: Return dictionary with results
Adjust the tests to use these results rather than the logged msgargs
value.  Also, modify the tests to check some items that should be
there without assuming a particular field order.

It would probably make sense to replace the msgargs tuple with a
dictionary instead, but that would involved updating the line
template, which doesn't seem worth the effort given that the plan is
to switch ls over to pyout.