CLN: Drop Emacs configuration line from file headers
These lines are visual noise for those that don't use Emacs and can be
an annoyance for those that do.  They configure py-indent-offset, but
this variable is from a third-party python mode.  Emacs users that
don't use this mode get prompted about possibly setting an unsafe
value and need to adjust their configuration to avoid being prompted
about this in the future.

_If_ we want to provide developers using Emacs with configuration for
this project, the better place to do it would be a top-level
.dir-locals.el.  But our settings at this point are typical for a
python project and match the default settings for Emacs's builtin
python.el.