ENH: job templates: Add some small files to git rather than annex
The main one we care about is status because we'll use 'git cat-file'
to get the status from the ref when the job is no longer in the
working tree.

This approach assumes that "git" is somewhere in PATH (rather than
just bundled with datalad), but the templates and orchestrators
already make this assumption elsewhere.