ENH: datalad orchestrators: Support retrieving status out-of-tree
The job metadata, including status information, is tracked under
.reproman/jobs/RESOURCE/JOBID/.  For DataLad orchestrators, we commit
these files on the remote after the run.  But when we check the status
file at a later time, we can't assume it is in the working tree.  In a
subsequent run, the remote's HEAD may have changed to match the local
repository's HEAD that didn't have the remote HEAD as an ancestor.
This can easily happen if the user executes another job before
fetching the last one.

Now that we store the status file in git, retrieve the status with
cat-file if we don't find it in the working tree.