ENH: orchestrators: Restore old HEAD when injecting run results
When we bring in the results and create a run commit, we do this on
the local dataset's HEAD.  But HEAD may have changed since we
submitted the job.  If we're not on the original HEAD, temporarily
restore it when injecting the results and let the user know how they
can bring the results into their current branch.

Point to the results commit with refs/reproman/JOBID.  (This is the
same refname we use remotely, though, in the case of the run
orchestrator, these aren't the same commit.)  We don't need to create
this ref; we could just tell the hash they should merge.  But the ref
makes it easier to find the results at a later time.  It also prevents
an unmerged results commit from being garbage collected as an
unreachable object.  And, for consistency, create the ref even when we
don't have to move HEAD.