TST: orchestrators: Reduce datatset creation
Add two fixtures, a module-scoped one that creates a dataset and a
function-scoped one that resets master to the original commit.  No
other clean-up is done, so it is quite easy to leak state, but
resetting master is sufficient for our current uses.  Performance-wise
it doesn't make much of a difference for the job submission tests,
which spend a good amount of time waiting for jobs, but it will make a
difference in tests for the upcoming head_at helper.