ENH: orchestrators: Fetch job ref in datalad-pair's fetch()
This ref is useful for temporarily moving HEAD when fetching the
outputs and for letting the user know where the results are if they
aren't merged into master by the update call.  It also means that some
trace of the job exists in the local dataset even if the update call
fails.