ENH: orchestrators: Move HEAD in datalad-pair's fetch
After the 'datalad update' call, the local HEAD will only match the
remote dataset's HEAD at the end of the run if there have been no
other commits in the local repository and no subsequent jobs executed
from this dataset before the run is fetched.  Before retrieving the
outputs, make sure the working tree matches the remote's at the time
the job results were committed.  Otherwise the files may not exist or
the content may have changed.

Unfortunately the ssh bits of fetch() (and also prepare_remote()) are
still untested.  We should adjust tests to either work with localhost
or a container.  Also, there are plans to make datalad's
create-sibling work locally, in which case much of this code could be
tested with the "local" setup.