ENH: orchestrators: Notify user if 'update' doesn't merge job's commit
The datalad-pair orchestrator relies on 'datalad update' to pull the
commit for the job from the remote, but it's possible for the merge
from the remote dataset to not include the job commit.  This can
happen when a job is submitted before fetching the results of the last
one.  If the job commit isn't part of the current branch, provide
instructions on how to incorporate the results into the current
branch.