TST: Improve readability of pycache_source test
The IDs of the parametrize test cases are in a separate list from the
values.  Given the number of test cases, these IDs aren't very helpful
as descriptions to a reader because it is hard to match an ID to a
test case.  Store the IDs alongside the test values to make this
clearer.