DOC: run: Remove meta_directory from listed job parameters
We hard code this as a subdirectory of working_directory and don't
expose it via the job spec.  Let's not change that until there is a
strong case made for doing so because knowing that the meta_directory
is a subdirectory of working_directory allows us to reuse the
"working_directory -> local_directory" mapping.