ENH: orchestrators: Configure .reproman/jobs/.gitattributes
When we run the job on the remote dataset, the job template 'git add's
status and idmap so that they are tracked in git.  But the local
dataset for 'datalad run'-based orchestrators inconsistently puts
these files under annex because the 'datalad add' call is done on the
local machine.

To consistently track these files in git, configure .gitattributes
under .reproman/jobs rather than using 'git add' in the job template.