BF: orchestrators: Push ref when creating sibling on local shell
This is consistent with the "already exists" branch of the code, is
useful for debugging, and means we get an "assert clean" check before
the push.

Though none of that matters too much.  The local shell bits of
datalad-pair's prepare_remote() exist at this point for local testing
and will be unnecessary once DataLad's create-sibling supports setting
up a local sibling.