ENH: ls: Use pyout to render table
The main benefit of using pyout here is that we can query statuses
asynchronously, but it also allows us to style the output in more
complex ways (automatic truncation, coloring by regexp, ...) with less
custom code than we'd need otherwise.

This switch to pyout should not change how we interpret or update the
resource statuses, but the code for interacting with the resources is
a little more complicated because, when --refresh is given, we access
the resources in functions that are called asynchronously.  And to
update the resource statuses, we need to look at the values after the
asynchronous calls have returned.

Closes #318.