TST: ls: Adjust return structure to mirror pyout
Ls.__call__() returns a dictionary with results to ease testing.
ls.py will start using pyout in the next commit.  Once that happens,
it makes sense to drop the custom result dictionary because the
Tabular object already contains the same information and provides
access to the values _after_ the asynchronous functions.

But before we do that, convert the result dictionary to a structure
that more closely matches the Tabular getitem interface so that we can
switch to pyout *without* modifying the actual tests.