ENH: ssh: Include resource name in password prompt
When the inventory contains a password-protected resource, calling
`niceman ls --refresh` will display a "Password: " prompt _before_
displaying the line for the resource.  Add the resource name to the
prompt so that users know what they are entering a password for.

For operations that work on a single resource (e.g., login), this
added information isn't useful, but it also doesn't hurt.

Re: #400