NF: run: Support specification of batch commands via yaml file
The orchestrator has been wired up to generate an array of commands
based on records.  For example, with

  [{"name": "foo"},
   {"name": "bar"}]

and a command "touch {p[name]}" would expand to two concurrent calls

  touch foo
  touch bar

Add an option to run() that allows the caller to point to a file that
these records.  This could be a yaml/json file or it could be a python
file that generates the record, but let's go with a yaml file for now
because it's a simple step from a python file to dumping a yaml file
and then python isn't forced on the user.

The next commit will add an alternative option that allows specifying
simple combinations via the command line without the need for a yaml
file.