BF: orchestrator: Fix target commit checkout for shell resource
When preparing a dataset in the local shell [*], the target checkout
occurs _before_ the push, but that of course fails when the target
needs to be sent by the push (i.e. if a future run occurs on a commit
that didn't exist when the remote was originally created).

Check out the target _after_ the push.

[*] create_sibling doesn't work on the local machine, so the
    local-shell-specific method is meant as a temporary workaround and
    exists mostly for testing.