BF: orchestrators: Fix plain copy for inputs in subdirectory
For a plain remote, we use the base name of the input for the
destination, which of course fails for inputs under a subdirectory.
Instead use the path relative to the local directory.

Note that the path handling here is a bit wonky because we use op.join
even though it is for a destination on the remote.  But that is a
problem shared by all our resource transfer methods and something that
those methods should likely handle.