ENH: jobs: Keep information on failed sub jobs
Once we support concurrent jobs, we need a way for each subjob to
indicate that it failed.  Teach failing jobs to touch a file in the
"METADIR/failed/".  We could instead check the content of status.N,
but that could mean checking the contents of a large number of files
just to obtain the status of (hopefully) a small number of failures.

Update log_failed() to look at this directory rather than .status,
which won't be meaningful for concurrent jobs.