TST: travis: Remove 'route' commands for NONETWORK run
The REPROMAN_TESTS_NONETWORK run is hanging on the 'route add'
command (see gh-420).  The reason for this isn't clear, but let's
follow DataLad (where this setup is inherited from) and remove these
commands.  See DataLad's 64a4dd7aa (Remove messing with the route
table -- not sure how/if it works currently, 2016-09-15).