ENH: VenvEnvironment: Add system_site_packages attribute
VenvPackage has a local attribute that is set to true if the package
is a local package (as classified by 'pip list --local'), so if
someone wanted to determine whether --system-site-packages is needed
to create the current environment, they could inspect VenvEnvironment
for any packages without local=True.  But this is tedious to do by
visual inspection, so add a system_site_packages attribute to
VenvEnvironment that explicitly marks whether the environment has
non-local packages.

Note that a system_site_packages value of False isn't a guarantee that
--system-site-packages was _not_ used: it may have been used but no
system site packages were installed at the time of tracing.

Closes #428.